refactor: rename minio_archiver module to s3_archiver (cosmetic)
Naming/log-string only — no behavior change, no logic change.
Class S3CompatibleArchiver is unchanged.

- git mv minio_archiver.py → s3_archiver.py
- git mv test_minio_archiver.py → test_s3_archiver.py
- Update import in archiver.py and test_archiver.py
- Reword 7 log strings/docstrings/comments: "minio …" → neutral
  object-storage terms
- README: add one sentence noting S3-compatible stores
  (Wasabi, R2, B2, Ceph, DO Spaces) work under type: s3 or type: minio
